+++ Bouy-Firmware/src/sensors/tdssensor.cpp
@@ -7,7 +7,7 @@
 #include "hardwareLayout.h"
 #include "sensor.h"
 
-TDSSensor::TDSSensor(uint32_t sensor_id, uint32_t type_id) : Sensor(sensor_id, type_id) {}
+TDSSensor::TDSSensor(uint16_t sensor_id, uint16_t type_id) : Sensor(sensor_id, type_id) {}
 
 std::vector<Value> TDSSensor::sampleValues() {
   double adcValue = analogRead(TDS_PIN);
